move update to git module
