+++ pr_cli/gh.py
@@ -20,7 +20,7 @@ def is_current_branch_updated(user):
 def current_pr(user, base='master'):
     repo = current_repo(user)
     branch = git.current_branch()
-    pulls = repo.get_pulls(base=base,head=branch, state='open')
+    pulls = repo.get_pulls(base=base, head=branch, state='open')
     try:
         return pulls[0]
     except IndexError:
